feat(metro-resolver-symlinks): add experimental `oxc-resolver`
