+++ .changeset/tasty-rules-check.md
@@ -1,0 +1,5 @@
+---
+"@rnx-kit/metro-resolver-symlinks": patch
+---
+
+Added experimental support for `oxc-resolver` as an option
