+++ packages/metro-resolver-symlinks/src/remappers/remapImportPath.ts
@@ -5,7 +5,7 @@ import {
 } from "@rnx-kit/tools-node";
 import type { PackageModuleRef } from "@rnx-kit/tools-node/module";
 import { expandPlatformExtensions } from "@rnx-kit/tools-react-native/platform";
-import * as path from "path";
+import * as path from "node:path";
 import type { ModuleResolver, ResolutionContextCompat } from "../types";
 
 type Resolver = (fromDir: string, moduleId: string) => string;
