+++ packages/metro-resolver-symlinks/src/utils/package.ts
