+++ packages/metro-resolver-symlinks/test/remappers/remapReactNative.test.ts
@@ -1,6 +1,5 @@
-import * as path from "node:path";
-import { remapReactNativeModule, resolveModulePath } from "../src/resolver";
-import { useFixture } from "./fixtures";
+import { remapReactNativeModule } from "../../src/remappers/remapReactNative";
+import { useFixture } from "../fixtures";
 
 const AVAILABLE_PLATFORMS = {
   macos: "react-native-macos",
@@ -49,36 +48,5 @@ describe("remapReactNativeModule", () => {
         `${npmPackage}/index`
       );
     });
-  });
-});
-
-describe("resolveModulePath", () => {
-  function makeContext(originModulePath: string) {
-    return { originModulePath };
-  }
-
-  test("returns absolute/relative modules as is", () => {
-    expect(resolveModulePath(makeContext(""), "./terminator", "")).toBe(
-      "./terminator"
-    );
-    expect(resolveModulePath(makeContext(""), "/terminator", "")).toBe(
-      "/terminator"
-    );
-  });
-
-  test("resolves module path relative to requester", () => {
-    const p = useFixture("duplicates");
-    expect(resolveModulePath(makeContext(p), "react-native", "")).toBe(
-      `.${path.sep}${path.join("duplicates", "node_modules", "react-native")}`
-    );
-    expect(
-      resolveModulePath(
-        makeContext(path.join(p, "node_modules", "terminator")),
-        "react-native",
-        ""
-      )
-    ).toBe(
-      `.${path.sep}${path.join("terminator", "node_modules", "react-native")}`
-    );
   });
 });
